modularizing(helpers/matchAcademicType.js)
